clean: small fixes

+++ src/nexus_trade/config/strategy.py
@@ -38,8 +38,8 @@ class TradingHoursConfig(FrozenModel):
 class RiskConfig(FrozenModel):
     max_positions: int = Field(default=1, gt=0)
     max_trades: int = Field(default=1, gt=0)
-    max_spread_points: int = Field(default=100, ge=0)
-    max_slippage_points: int = Field(default=100, ge=0)
+    max_spread_points: float = Field(default=100, ge=0)
+    max_slippage_points: float = Field(default=100, ge=0)
 
 
 class BaseStrategyParams(FrozenModel):
